+++ hera_sim/foregrounds.py
@@ -149,7 +149,7 @@ class DiffuseForeground(Foreground):
 
         # resample the sky temperature model
         Tsky = Tsky_mdl(lsts=lsts, freqs=freqs)  # K
-        vis = np.asarray(Tsky / utils.Jy2T(freqs, omega_p), np.complex)
+        vis = np.asarray(Tsky / utils.jansky_to_kelvin(freqs, omega_p), np.complex)
 
         if np.isclose(np.linalg.norm(bl_vec), 0):
             return vis