+++ hera_sim/noise.py
@@ -84,7 +84,7 @@ class ThermalNoise(Noise):
 
         # get the sky temperature; use an autocorrelation if provided
         if autovis is not None and not np.all(np.isclose(autovis, 0)):
-            Tsky = autovis * utils.Jy2T(freqs, omega_p).reshape(1, -1)
+            Tsky = autovis * utils.jansky_to_kelvin(freqs, omega_p).reshape(1, -1)
         else:
             Tsky = self.resample_Tsky(lsts, freqs, Tsky_mdl=Tsky_mdl)
 
@@ -97,7 +97,7 @@ class ThermalNoise(Noise):
 
         # convert vis to Jy
         # XXX why the reshape?
-        vis /= utils.Jy2T(freqs, omega_p).reshape(1, -1)
+        vis /= utils.jansky_to_kelvin(freqs, omega_p).reshape(1, -1)
 
         # make it noisy
         return utils.gen_white_noise(size=vis.shape) * vis